various minor cleanup fixes
git-svn-id: svn://192.168.1.10/public/CoffeeMud@20804 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Libraries/Achievements.java
@@ -1080,7 +1080,7 @@ public class Achievements extends StdLibrary implements AchievementLibrary
 								if((killed != null)
 								&&((npcMask==null)||(CMLib.masking().maskCheck(npcMask, killed, true)))
 								&&((playerMask==null)||(CMLib.masking().maskCheck(playerMask, mob, true)))
-								&&((gSize < 2)||(grpSize>=gSize)))
+								&&((gSize < 2)||(grpSize.intValue()>=gSize)))
 								{
 									count+=bumpNum;
 									return true;
